+++ src/arco_era5/source_data.py
@@ -339,7 +339,9 @@ CUMULATIVE_VARIABLES = ("total_precipitation",
 PRESSURE_LEVELS_GROUPS = immutabledict.immutabledict({
     "weatherbench_13":
         (50, 100, 150, 200, 250, 300, 400, 500, 600, 700, 850, 925, 1000),
-    "full_37": (1, 2)
+    "full_37": (1, 2, 3, 5, 7, 10, 20, 30, 50, 70, 100, 125, 150, 175, 200, 225,
+                250, 300, 350, 400, 450, 500, 550, 600, 650, 700, 750, 775, 800,
+                825, 850, 875, 900, 925, 950, 975, 1000)
 })
 
 # ERA5 uses "geopotential" for file names for both the static "geopotential"
